BIGTOP-1986. Extract location dataset from BigPetStore data generator
